+++ CMakeLists.txt
@@ -23,7 +23,7 @@ add_compile_definitions(DUCKDB_AWS_GIT_SHA="${DUCKDB_AWS_GIT_SHA}")
 
 set(EXTENSION_SOURCES src/aws_extension.cpp src/aws_secret.cpp src/cloudformation_functions.cpp
                       src/redshift/redshift_utils.cpp src/redshift/redshift_storage.cpp
-                      src/utils/region_utils.cpp)
+                      src/utils/utils.cpp)
 add_library(${EXTENSION_NAME} STATIC ${EXTENSION_SOURCES})
 
 set(PARAMETERS "-warnings")
